+++ src/test/java/com/hazlewood/coffeetracker/beans/BeansRepositoryIntegrationTest.java
@@ -6,6 +6,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.test.context.ActiveProfiles;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -27,7 +28,7 @@ class BeansRepositoryIntegrationTest {
 
   @Test
   public void whenFindByName_ThenReturnEmployee() {
-    Beans ancoats = new Beans("Ancoats house blend", 500);
+    Beans ancoats = new Beans("Ancoats house blend", BigDecimal.valueOf(500));
     entityManager.persistAndFlush(ancoats);
 
     Beans found = repository.findByName("Ancoats house blend").get();
@@ -42,7 +43,7 @@ class BeansRepositoryIntegrationTest {
 
   @Test
   public void whenFindById_ThenReturnBeans() {
-    Beans ancoats = new Beans("Ancoats house blend", 500);
+    Beans ancoats = new Beans("Ancoats house blend", BigDecimal.valueOf(500));
     entityManager.persistAndFlush(ancoats);
 
     Beans found = repository.findById(ancoats.getId()).get();
@@ -57,9 +58,9 @@ class BeansRepositoryIntegrationTest {
 
   @Test
   public void givenSetOfBeans_WhenFindAll_ThenReturnAllBeans() {
-    var ancoats = new Beans("Ancoats house blend", 500);
-    var atkinsons = new Beans("Atkinsons house blend", 500);
-    var neighbourhood = new Beans("Neighbourhood house blend", 500);
+    var ancoats = new Beans("Ancoats house blend", BigDecimal.valueOf(500));
+    var atkinsons = new Beans("Atkinsons house blend", BigDecimal.valueOf(500));
+    var neighbourhood = new Beans("Neighbourhood house blend", BigDecimal.valueOf(500));
 
     entityManager.persist(ancoats);
     entityManager.persist(atkinsons);
